Create an example for the message passing function

+++ datasets/utils.py
@@ -12,6 +12,7 @@ from __future__ import print_function
 
 import networkx as nx
 import rdkit
+import torch
 
 __author__ = "Pau Riba, Anjan Dutta"
 __email__ = "priba@cvc.uab.cat, adutta@cvc.uab.cat"
@@ -38,7 +39,7 @@ def qm9_nodes(g, hydrogen=False):
         # If number hydrogen is used as a
         if hydrogen:
             h_t.append(d['num_h'])
-        h[n] = h_t
+        h[n] = torch.FloatTensor(h_t)
     return h
 
 
@@ -80,5 +81,5 @@ def qm9_edges(g, e_representation='chem_graph'):
             print('Incorrect Edge representation transform')
             quit()
         if e_t:
-            e[(n1, n2)] = e_t
+            e[(n1, n2)] = torch.FloatTensor(e_t)
     return g, e
